Add test base classes for tests in aggregation module
Otherwise, many tests that will be moved need to overwrite getPlugin(...) or nodePlugin(...)
method and supply the AggregationsPlugin.class as plugin class.

Added the following base classes:
* AggregationIntegTestCase for aggregation java integ tests.
* AggregationBuilderTestCase for testing aggregation builders.
* AggregationMultiBucketAggregationTestCase for testing multi bucket response classes.
  This also simplifies supplying `ParsedAggregation` parsers a little bit.
* AggregationTestCase for testing aggregators in this module.

Relates to #90283
